+++ Backend/src/main/java/app/entity/Container.java
@@ -16,7 +16,7 @@ public class Container {
 
     private float width;
 
-    private Integer capacity;
+    private float capacity;
 
     private String material;
 
@@ -56,11 +56,11 @@ public class Container {
         this.width = width;
     }
 
-    public Integer getCapacity() {
+    public float getCapacity() {
         return capacity;
     }
 
-    public void setCapacity(Integer capacity) {
+    public void setCapacity(float capacity) {
         this.capacity = capacity;
     }
