+++ Backend/src/main/java/app/model/BeerStockModel.java
@@ -2,5 +2,5 @@ package app.model;
 
 public class BeerStockModel {
     public Long id;
-    public Integer quantity;
+    public float quantity;
 }
