+++ Backend/src/main/java/app/model/ContainerModel.java
@@ -10,7 +10,7 @@ public class ContainerModel {
 
     private float width;
 
-    private Integer capacity;
+    private float capacity;
 
     private String material;
 
@@ -50,11 +50,11 @@ public class ContainerModel {
         this.width = width;
     }
 
-    public Integer getCapacity() {
+    public float getCapacity() {
         return capacity;
     }
 
-    public void setCapacity(Integer capacity) {
+    public void setCapacity(float capacity) {
         this.capacity = capacity;
     }
